penguin position fix

+++ include/Vec2.h
@@ -8,7 +8,7 @@ class Vec2 {
         float x, y;
 
         Vec2();
-        Vec2(int x, int y);
+        Vec2(float x, float y);
         Vec2 GetRotated(float theta);
         Vec2 operator+(const Vec2& vec);
         Vec2 operator-(const Vec2& vec);
