+++ src/Vec2.cpp
@@ -5,7 +5,7 @@ Vec2::Vec2 () {
     y = 0;
 }
 
-Vec2::Vec2 (int x, int y) {
+Vec2::Vec2 (float x, float y) {
     this->x = x;
     this->y = y;
 }
